Add files via upload
Basic number range check to see if a value is below 100, within 100 to 1000, within 1001 to 2000, or above 2000. 

I would imagine there is a way to do this in a lot less code. If anyone actually sees this please feel free to show me. 

Maybe I'll update it way later on and see some progress or improvements. For now it's what I could do!